Update .clang-format and add format_count.sh

+++ CodeTop/CodeTop016.cc
@@ -3,8 +3,8 @@
  * @link https://leetcode.cn/problems/number-of-islands/
  */
 
-#include <iostream>
 #include <functional>
+#include <iostream>
 #include <vector>
 
 using std::cout;
@@ -17,13 +17,12 @@ public:
         if (grid.empty()) return 0;
 
         int cnt = 0;
-        std::function<void (int, int)> dfs = [&] (int x, int y) {
-            if (x < 0 || x >= grid.size() ||
-                y < 0 || y >= grid[x].size() ||
+        std::function<void(int, int)> dfs = [&](int x, int y) {
+            if (x < 0 || static_cast<size_t>(x) >= grid.size() || y < 0 || static_cast<size_t>(y) >= grid[x].size() ||
                 grid[x][y] != '1') {
                 return;
             }
-            
+
             grid[x][y] = '*';
             dfs(x - 1, y);
             dfs(x + 1, y);
@@ -43,17 +42,14 @@ public:
     }
 };
 
-int main () {
+int main() {
     // vector<vector<char>> grid {
     //     {'1', '1', '0', '0', '0'},
     //     {'1', '1', '0', '0', '0'},
     //     {'0', '0', '1', '0', '0'},
     //     {'0', '0', '0', '1', '1'}};
 
-    vector<vector<char>> grid {
-        {'1', '1', '1'},
-        {'0', '1', '0'},
-        {'0', '1', '0'}};
+    vector<vector<char>> grid{{'1', '1', '1'}, {'0', '1', '0'}, {'0', '1', '0'}};
 
     Solution so;
     cout << so.numIslands(grid) << endl;
